# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  handlers.PasswordResetEmailRequest:
    properties:
      name:
        type: string
      reset_code:
        type: string
      to:
        type: string
    required:
    - name
    - reset_code
    - to
    type: object
...
  handlers.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  handlers.VerificationEmailRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      token:
        type: string
    required:
    - email
    - first_name
    - token
    type: object
  handlers.WelcomeEmailRequest:
    properties:
      name:
        type: string
      to:
        type: string
      verification_code:
        type: string
    required:
    - name
    - to
    - verification_code
    type: object
...
  notification.Notification:
    properties:
      action:
        type: string
      created_at:
        type: string
      data: {}
      entity:
        type: string
      entity_id:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      level:
        type: string
      message:
        type: string
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  services.EmailRequest:
    properties:
      bcc:
        items:
          type: string
        type: array
      body:
        type: string
      cc:
        items:
          type: string
        type: array
      is_html:
        type: boolean
      subject:
        type: string
      template_id:
        type: string
      template_vars:
        additionalProperties: true
        type: object
      to:
        items:
          type: string
        type: array
    required:
    - subject
    - to
    type: object
  services.EmailResponse:
    properties:
      message:
        type: string
      sent_at:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@forgecrud.com
    name: API Support
    url: http://www.forgecrud.com/support
  description: Central API documentation - For all ForgeCRUD microservices
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ForgeCRUD API
  version: "1.0"
paths:
  /api/notifications/email/password-reset:
    post:
      consumes:
      - application/json
      description: Send a password reset email with reset code using template
      parameters:
      - description: Password reset email request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.EmailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send password reset email
      tags:
      - email
  /api/notifications/email/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend verification email to user after creating new token
      parameters:
      - description: Resend verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Resend verification email
      tags:
      - email
  /api/notifications/email/send:
    post:
      consumes:
      - application/json
      description: Send an email through the notification service
      parameters:
      - description: Email request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/services.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.EmailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send email
      tags:
      - email
  /api/notifications/email/verification:
    post:
      consumes:
      - application/json
      description: Send email verification link to user
      parameters:
      - description: Verification email request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerificationEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send verification email
      tags:
      - email
  /api/notifications/email/welcome:
    post:
      consumes:
      - application/json
      description: Send a welcome email with verification code using template
      parameters:
      - description: Welcome email request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/handlers.WelcomeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.EmailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send welcome/verification email
      tags:
      - email
...
  /notifications:
    get:
      consumes:
      - application/json
      description: Get all notifications for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/notification.Notification'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all notifications
      tags:
      - notifications
    post:
      consumes:
      - application/json
      description: Create a new notification
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification.Notification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/notification.Notification'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create notification
      tags:
      - notifications
  /notifications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete notification
      tags:
      - notifications
    get:
      consumes:
      - application/json
      description: Get a specific notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.Notification'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get notification by ID
      tags:
      - notifications
  /notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.Notification'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...